Complete removal of test-client to embed into the native client
