+++ client.js
@@ -16,14 +16,19 @@ class OSCRelayClient {
         this.parameters = new Map();
 
         this.localOscPort = 9000;
-        this.vrchatSendPort = 9000;
-        this.vrchatReceivePort = 9001;
+        this.vrchatSendPort = 9001;
+        this.vrchatReceivePort = 9000;
 
         this.vrchatSender = new osc.Client('127.0.0.1', this.vrchatSendPort);
         this.setupOSCReceiver();
 
         this.startOSCQuery();
         this.setupKeyboardInput();
+
+        this.isTestMode = process.argv.includes('--test');
+        if (this.isTestMode) {
+            this.runTestMode();
+        }
     }
 
     setupOSCReceiver() {
@@ -47,7 +52,7 @@ class OSCRelayClient {
         } catch (err) {
             console.error('[Client] Failed to setup OSC receiver:', err);
             this.localOscPort++;
-            if (this.localOscPort < 9020) { // Try up to port 9020
+            if (this.localOscPort < 9020) {
                 this.setupOSCReceiver();
             }
         }
@@ -285,6 +290,53 @@ class OSCRelayClient {
         console.log('  [Ctrl+C] Exit application');
         console.log('===============================\n');
     }
+
+    async runTestMode() {
+        try {
+            console.log('[Test Mode] Setting up OSC and WebSocket...');
+            await this.connect();
+
+            console.log('[Test Mode] Testing parameter discovery...');
+            await new Promise(resolve => setTimeout(resolve, 2000));
+
+            const params = this.getAllParameters();
+            console.log('[Test Mode] Available parameters:', Object.keys(params).length);
+
+            this.onMessage(message => {
+                console.log('[Test Mode] Received OSC Message:', {
+                    timestamp: new Date().toISOString(),
+                    source: message.source || 'unknown',
+                    address: message.address,
+                    args: message.args
+                });
+            });
+
+            if (this.connected) {
+                const testCases = [
+                    { path: '/avatar/parameters/Voice', value: 0.75 },
+                    { path: '/avatar/parameters/VRCEmote', value: 1 },
+                ];
+
+                console.log('[Test Mode] Running test cases...');
+                for (const test of testCases) {
+                    await new Promise(resolve => setTimeout(resolve, 500));
+                    this.send(test.path, test.value);
+                    console.log(`[Test Mode] Sent: ${test.path} = ${test.value}`);
+                }
+            }
+        } catch (error) {
+            console.error('[Test Mode] Error:', error.message);
+            process.exit(1);
+        }
+    }
+}
+
+// Add direct execution support
+if (require.main === module) {
+    const SERVER_HOST = '57.128.188.155';
+    const SERVER_PORT = 4953;
+    
+    const client = new OSCRelayClient(`ws://${SERVER_HOST}:${SERVER_PORT}`);
 }
 
 module.exports = OSCRelayClient;
